+++ src/text/story.tsx
@@ -4,10 +4,21 @@ import { storiesOf } from '@storybook/react'
 import { withInfo } from '@storybook/addon-info'
 import { selectV2, text, withKnobs, boolean } from '@storybook/addon-knobs'
 
+import { color } from '_utils/branding'
 import Text, { TextTagType, TextDisplayType } from 'text'
 
 const stories = storiesOf('Text', module)
 stories.addDecorator(withKnobs)
+
+const c = Object.keys(color).reduce(
+  (acc, key) => ({
+    ...acc,
+    [key]: color[key],
+  }),
+  {},
+)
+
+console.log(c)
 
 stories.add(
   'basic',
@@ -15,6 +26,7 @@ stories.add(
     <Text
       display={selectV2('Display', TextDisplayType, TextDisplayType.BODY)}
       tag={selectV2('Tag', TextTagType, TextTagType.SPAN)}
+      textColor={selectV2('Color', c, 'primary')}
       newlineToBr={boolean('newlineToBr', true)}
     >
       {text('Text', 'This is an example')}